Use static clusters instead of EDS
It appears that when a cluster is updated without a following endpoint
update a cluster can lose its endpoint configuration.

I attempted to update endpoints on each cluster change to prevent
this but it was still possible for the cluster update to happen after
the endpoint update.

This appears related to https://github.com/envoyproxy/envoy/issues/13009

By using static clusters the endpoints are included in the cluster update.
In testing this appears to resolve the issue where endpoints disappear
since we no longer rely on EDS.